chore(base): Move `Room::*room_call*` methods in the new `call` module.
This patch moves the `Room::has_active_room_call` and
`Room::active_room_call_participants` methods into the new `call`
module. This patch also moves the associated tests in this new module.
